requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-OAuthlib==0.9.5
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-WTF==0.14.2
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mongoengine==0.17.0
oauth2client==4.1.3
oauthlib==2.1.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pymongo==3.7.2
python-dateutil==2.8.0
requests==2.21.0
requests-oauthlib==1.1.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.3
WTForms==2.2.1
